requirements.txt
pandas
matplotlib
sqlalchemy
seaborn
scipy
datetime
numpy
statsmodels
PyYAML
scikit-learn
xgboost
pytest
category_encoders